melhoria no controller para excluir um usuaria.

+++ src/controllers/user.controller.ts
@@ -1,6 +1,7 @@
 import type { Request, Response } from "express";
 import { UserService } from "../services/user.service.js";
 import type { AuthRequest } from "../middleware/auth.middleware.js";
+import bcrypt from "bcryptjs";
 
 export const UserController = {
   getUser: async (req: AuthRequest, res: Response) => {
@@ -24,12 +25,21 @@ export const UserController = {
   },
 
   deleteUserById: async (req: AuthRequest, res: Response) => {
+    const id = req.user?.userId as number;
+    const { pass } = req.body
     try {
-      const id = req.user?.userId as number;
+      const user = await UserService.findById(id)
+      if(!user) {
+        throw new Error("token invalido.")
+      }
+      const authPassword = await bcrypt.compare(pass, user.password)
+      if(!authPassword) {
+        throw new Error("Você não tem permição para excluir esse usuário.")
+      }
       const deleteUser = await UserService.deleteById(id);
-      res.status(200).json(deleteUser);
+      res.status(200).json({ message: "Usuário deletado com sucesso."});
     } catch (error) {
-      res.status(500).json({ message: "Não foi possivel deletar o usuário." });
+      res.status(500).json({ message: error });
     }
   },
